chore: clean up policy router
