+++ packages/api/src/router/policy/deny-window.ts
@@ -1,0 +1,78 @@
+import { z } from "zod";
+
+import { eq, takeFirst } from "@ctrlplane/db";
+import {
+  createPolicyRuleDenyWindow,
+  policyRuleDenyWindow,
+  updatePolicyRuleDenyWindow,
+} from "@ctrlplane/db/schema";
+import { Permission } from "@ctrlplane/validators/auth";
+
+import { createTRPCRouter, protectedProcedure } from "../../trpc";
+
+export const policyDenyWindowRouter = createTRPCRouter({
+  create: protectedProcedure
+    .meta({
+      authorizationCheck: ({ canUser, input }) =>
+        canUser
+          .perform(Permission.PolicyCreate)
+          .on({ type: "policy", id: input.policyId }),
+    })
+    .input(createPolicyRuleDenyWindow)
+    .mutation(({ ctx, input }) => {
+      return ctx.db
+        .insert(policyRuleDenyWindow)
+        .values(input)
+        .returning()
+        .then(takeFirst);
+    }),
+
+  update: protectedProcedure
+    .meta({
+      authorizationCheck: async ({ canUser, input, ctx }) => {
+        const denyWindow = await ctx.db
+          .select()
+          .from(policyRuleDenyWindow)
+          .where(eq(policyRuleDenyWindow.id, input.id))
+          .then(takeFirst);
+
+        return canUser
+          .perform(Permission.PolicyUpdate)
+          .on({ type: "policy", id: denyWindow.policyId });
+      },
+    })
+    .input(
+      z.object({ id: z.string().uuid(), data: updatePolicyRuleDenyWindow }),
+    )
+    .mutation(({ ctx, input }) => {
+      return ctx.db
+        .update(policyRuleDenyWindow)
+        .set(input.data)
+        .where(eq(policyRuleDenyWindow.id, input.id))
+        .returning()
+        .then(takeFirst);
+    }),
+
+  delete: protectedProcedure
+    .meta({
+      authorizationCheck: async ({ canUser, input, ctx }) => {
+        const denyWindow = await ctx.db
+          .select()
+          .from(policyRuleDenyWindow)
+          .where(eq(policyRuleDenyWindow.id, input))
+          .then(takeFirst);
+
+        return canUser
+          .perform(Permission.PolicyDelete)
+          .on({ type: "policy", id: denyWindow.policyId });
+      },
+    })
+    .input(z.string().uuid())
+    .mutation(({ ctx, input }) =>
+      ctx.db
+        .delete(policyRuleDenyWindow)
+        .where(eq(policyRuleDenyWindow.id, input))
+        .returning()
+        .then(takeFirst),
+    ),
+});
