+++ packages/api/src/router/policy/target.ts
@@ -1,0 +1,72 @@
+import { z } from "zod";
+
+import { eq, takeFirst } from "@ctrlplane/db";
+import {
+  createPolicyTarget,
+  policyTarget,
+  updatePolicyTarget,
+} from "@ctrlplane/db/schema";
+import { Permission } from "@ctrlplane/validators/auth";
+
+import { createTRPCRouter, protectedProcedure } from "../../trpc";
+
+export const policyTargetRouter = createTRPCRouter({
+  create: protectedProcedure
+    .meta({
+      authorizationCheck: ({ canUser, input }) =>
+        canUser
+          .perform(Permission.PolicyCreate)
+          .on({ type: "policy", id: input.policyId }),
+    })
+    .input(createPolicyTarget)
+    .mutation(({ ctx, input }) =>
+      ctx.db.insert(policyTarget).values(input).returning().then(takeFirst),
+    ),
+
+  update: protectedProcedure
+    .meta({
+      authorizationCheck: async ({ canUser, input, ctx }) => {
+        const target = await ctx.db
+          .select()
+          .from(policyTarget)
+          .where(eq(policyTarget.id, input.id))
+          .then(takeFirst);
+
+        return canUser
+          .perform(Permission.PolicyUpdate)
+          .on({ type: "policy", id: target.policyId });
+      },
+    })
+    .input(z.object({ id: z.string().uuid(), data: updatePolicyTarget }))
+    .mutation(({ ctx, input }) =>
+      ctx.db
+        .update(policyTarget)
+        .set(input.data)
+        .where(eq(policyTarget.id, input.id))
+        .returning()
+        .then(takeFirst),
+    ),
+
+  delete: protectedProcedure
+    .meta({
+      authorizationCheck: async ({ canUser, input, ctx }) => {
+        const target = await ctx.db
+          .select()
+          .from(policyTarget)
+          .where(eq(policyTarget.id, input))
+          .then(takeFirst);
+
+        return canUser
+          .perform(Permission.PolicyDelete)
+          .on({ type: "policy", id: target.policyId });
+      },
+    })
+    .input(z.string().uuid())
+    .mutation(({ ctx, input }) =>
+      ctx.db
+        .delete(policyTarget)
+        .where(eq(policyTarget.id, input))
+        .returning()
+        .then(takeFirst),
+    ),
+});
